+++ src/app/features/home/home.component.ts
@@ -1,18 +1,20 @@
 import { Component, WritableSignal, signal } from '@angular/core';
-import { map } from 'rxjs';
+import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
 
 import { Amiibo } from '../../models/amiibo';
+import { AmiiboFilter } from '../../models/amiibo-filter';
+import { AmiiboListFilter } from '../../models/amiibo-list-filter';
 
 import { AmiiboService } from '../../services/amiibo.service';
 
-import { CommonModule } from '@angular/common';
 import { AmiiboItemComponent } from '../amiibo/components/amiibo-item/amiibo-item.component';
 import { AmiiboListComponent } from '../amiibo/components/amiibo-list/amiibo-list.component';
 
 @Component({
   selector: 'app-home',
   standalone: true,
-  imports: [CommonModule, AmiiboItemComponent, AmiiboListComponent],
+  imports: [CommonModule, FormsModule, AmiiboItemComponent, AmiiboListComponent],
   templateUrl: './home.component.html',
   styleUrl: './home.component.scss'
 })
@@ -20,7 +22,11 @@ export class HomeComponent {
 
   public totalAmiibos: WritableSignal<number> = signal(0);
   public amiibos: WritableSignal<Array<Amiibo>> = signal([]);
+  public amiiboSeries: WritableSignal<Array<AmiiboFilter>> = signal([]);
+  public amiiboTypes: WritableSignal<Array<AmiiboFilter>> = signal([]);
+  public amiiboCharacters: WritableSignal<Array<AmiiboFilter>> = signal([]);
   public isLoading: WritableSignal<boolean> = signal(false);
+  public filter: WritableSignal<AmiiboListFilter> = signal({ });
 
   constructor(private _amiiboService: AmiiboService) { }
 
@@ -28,4 +34,38 @@ export class HomeComponent {
     this.totalAmiibos.set(amiiboList.amiibo.length);
     this.amiibos.set(amiiboList.amiibo);
   });
+
+  public amiiboSeries$ = this._amiiboService.getAmiiboSeries().subscribe(amiiboSeriesList => {
+    this.amiiboSeries.set(amiiboSeriesList.amiibo);
+  });
+
+  public amiiboTypes$ = this._amiiboService.getAmiiboTypes().subscribe(amiiboTypesList => {
+    this.amiiboTypes.set(amiiboTypesList.amiibo);
+  });
+
+  public amiiboCharacters$ = this._amiiboService.getAmiiboCharacters().subscribe(amiiboCharactersList => {
+    this.amiiboCharacters.set(amiiboCharactersList.amiibo);
+  });
+
+  public updateAmiiboSeries(gameSeries: string) {
+    this.filter.set({...this.filter(), gameSeries });
+    this._updateList();
+  }
+
+  public updateAmiiboType(type: string) {
+    this.filter.set({...this.filter(), type });
+    this._updateList();
+  }
+
+  public updateAmiiboCharacter(character: string) {
+    this.filter.set({...this.filter(), character });
+    this._updateList();
+  }
+
+  private _updateList(): void {
+    this._amiiboService.getAmiibos(this.filter()).subscribe(amiiboList => {
+      this.totalAmiibos.set(amiiboList.amiibo.length);
+      this.amiibos.set(amiiboList.amiibo);
+    });
+  }
 }