+++ src/app/models/amiibo-filter.ts
@@ -1,0 +1,4 @@
+export interface AmiiboFilter {
+    key: string;
+    name: string;
+}