+++ src/app/models/amiibo-game.ts
@@ -1,0 +1,7 @@
+import { AmiiboUsage } from "./amiibo-usage";
+
+export interface AmiiboGame {
+  amiiboUsage: AmiiboUsage;
+  gameID: Array<string>;
+  gameName: string;
+}