+++ src/app/models/amiibo-list-filter.ts
@@ -1,0 +1,5 @@
+export interface AmiiboListFilter {
+  gameSeries?: string;
+  type?: string;
+  character?: string;
+}