+++ src/app/models/amiibo-usage.ts
@@ -1,0 +1,4 @@
+export interface AmiiboUsage {
+  Usage: string;
+  write: string;
+}