+++ src/app/models/data/amiibo-filter.response.ts
@@ -1,0 +1,5 @@
+import { AmiiboFilter } from "../amiibo-filter";
+
+export interface AmiiboFilterResponse {
+  amiibo: Array<AmiiboFilter>;
+}